solution/backend/requirements.txt
keras==2.12.0
numpy==1.23.5
pandas==2.2.1
plotly==5.19.0
scikit-learn==1.4.1.post1
# tensorboard==2.14.1
# tensorflow==2.14.0
# developer
pytest==8.0.2
nbformat==5.9.2
fastapi==0.110.0
uvicorn==0.27.1
PyYAML==6.0.1
pydantic==2.3.0
scikeras[tensorflow]